pyproject.toml
[tool.poetry]
name = "ltlf2dfa"
version = "1.0.2"
description = "From LTLf/PPLTL to Deterministic Finite-state Automata (DFA)"
authors = ["Francesco Fuggitti <francesco.fuggitti@gmail.com>"]
license = "LGPLv3+ license"
readme = "README.md"
homepage = "https://whitemech.github.io/ltlf2dfa"
repository = "https://github.com/whitemech/ltlf2dfa.git"
documentation = "https://whitemech.github.io/ltlf2dfa"
keywords = [
    "linear-time temporal logic on finite traces",
    "pure-past linear temporal logic",
    "deterministic finite-state automata",
]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Education',
    'License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
include = []

[tool.poetry.scripts]
ltlf2dfa = 'ltlf2dfa.__main__:cli'

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/whitemech/ltlf2dfa/issues"
"Pull Requests" = "https://github.com/whitemech/ltlf2dfa/pulls"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
click = "^8.1.3"
lark-parser = "^0.12.0"
sympy = "^1.12"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.3.0"
click = "^8.1.3"
codecov = "^2.1.13"
darglint = "^1.8.1"
flake8 = "^6.0.0"
flake8-bugbear = "^23.6.5"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-isort = "^6.0.0"
hypothesis = "^6.76.0"
hypothesis-pytest = "^0.19.0"
ipython = "^8.12.2"
isort = "^5.12.0"
jupyter = "^1.0.0"
markdown = "^3.3.4"
markdown-include = "^0.8.1"
mistune = "^2.0.5"
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.15"
mknotebooks = "^0.7.1"
mypy = "^1.3.0"
packaging = "^23.0"
pre-commit = "^3.3.2"
pylint = "^2.17.4"
pymdown-extensions = "^10.0.1"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
pytest-lazy-fixture = "^0.6.3"
pytest-randomly = "^3.12.0"
safety = "^2.3.5"
toml = "^0.10.2"
tox = "^4.4.12"
twine = "^4.0.2"
types-requests = "^2.31.0.1"
types-setuptools = "^67.8.0.0"
types-toml = "^0.10.8.6"
vulture = "^2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
